frontend/glaweb/requirements.txt
Django>=4.0.0
faiss-cpu>=1.7.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
numpy>=1.23.5,<2.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0